refactor: moving TripInfoContext from 'src\components\EditPage\temp_data\trip_reducer.jsx' to 'src\contexts\TripInfoContext.jsx'
