+++ src/components/EditPage/Panel/PanelBody/CRUD/CardBtnPopper.jsx
@@ -9,11 +9,13 @@ import Modal from '@mui/material/Modal';
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
 import { ClickAwayListener } from '@mui/base/ClickAwayListener';
 import { patchDestinations, deleteDestinations } from '@/api/editPage.js';
-
 import {
   tripInfo_actions,
   useTripInfo,
   useTripInfoDispatch,
+} from '@/contexts/TripInfoContext';
+
+import {
   routes_actions,
   useRoutesDispatch,
 } from '../../../temp_data/trip_reducer';
