+++ src/components/EditPage/Panel/PanelBody/Itinerary/DayItineraryContent.jsx
@@ -5,9 +5,9 @@ import Typography from '@mui/material/Typography';
 import DestinationItem from './DestinationItem';
 import TransportationItem from './TransportationItem';
 import { getRoutes, postRoutes } from '@/api/editPage.js';
+import { useTripInfo } from '@/contexts/TripInfoContext';
 
 import {
-  useTripInfo,
   routes_actions,
   useRoutes,
   useRoutesDispatch,
